+++ kibot/misc.py
@@ -89,6 +89,7 @@ IFILT_ROT_FOOTPRINT = '_rot_footprint'
 IFILT_KICOST_RENAME = '_kicost_rename'
 IFILT_KICOST_DNP = '_kicost_dnp'
 IFILT_EXPAND_TEXT_VARS = '_expand_text_vars'
+IFILT_DATASHEET_LINK = '_datasheet_link'
 # KiCad 5 GUI values for the attribute
 UI_THT = 0       # 1 for KiCad 6
 UI_SMD = 1       # 2 for KiCad 6